app/config.py

import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
    MODEL_NAME = "gpt-4o-mini"
    MAX_LOG_LINES = 200
    TEMPERATURE = 0.2

settings = Settings()
