requirements.txt

fastapi
uvicorn
python-multipart
jinja2
openai
python-dotenv
pydantic
